packages/service-integration/requirements/_base.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements/_base.txt --strip-extras requirements/_base.in
#
attrs==22.1.0
    # via
    #   jsonschema
    #   pytest
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   -r requirements/_base.in
    #   typer
colorama==0.4.6
    # via typer
commonmark==0.9.1
    # via rich
dnspython==2.2.1
    # via email-validator
docker==6.0.1
    # via -r requirements/_base.in
email-validator==1.3.0
    # via pydantic
exceptiongroup==1.0.1
    # via pytest
idna==3.4
    # via
    #   email-validator
    #   requests
iniconfig==1.1.1
    # via pytest
jsonschema==4.17.0
    # via
    #   -r requirements/../../../packages/models-library/requirements/_base.in
    #   -r requirements/_base.in
packaging==21.3
    # via
    #   docker
    #   pytest
pluggy==1.0.0
    # via pytest
pydantic==1.10.2
    # via
    #   -c requirements/../../../packages/models-library/requirements/../../../requirements/constraints.txt
    #   -c requirements/../../../requirements/constraints.txt
    #   -r requirements/../../../packages/models-library/requirements/_base.in
pygments==2.13.0
    # via rich
pyparsing==3.0.9
    # via packaging
pyrsistent==0.19.2
    # via jsonschema
pytest==7.2.0
    # via -r requirements/_base.in
pyyaml==6.0
    # via
    #   -c requirements/../../../packages/models-library/requirements/../../../requirements/constraints.txt
    #   -c requirements/../../../requirements/constraints.txt
    #   -r requirements/_base.in
requests==2.28.1
    # via docker
rich==12.6.0
    # via typer
shellingham==1.5.0
    # via typer
tomli==2.0.1
    # via pytest
typer==0.7.0
    # via -r requirements/_base.in
typing-extensions==4.4.0
    # via pydantic
urllib3==1.26.12
    # via
    #   -c requirements/../../../packages/models-library/requirements/../../../requirements/constraints.txt
    #   -c requirements/../../../requirements/constraints.txt
    #   docker
    #   requests
websocket-client==1.4.2
    # via docker
